+++ spec/features/early_career_payments_claim_spec.rb
@@ -194,7 +194,13 @@ RSpec.feature "Teacher Early Career Payments claims" do
     click_on "Continue"
 
     expect(claim.reload.student_loan_country).to eql("england")
-    # TODO [PAGE 32] - How many higher education courses did you take a student loan out for
+    # [PAGE 32] - How many higher education courses did you take a student loan out for
+    expect(page).to have_text(I18n.t("questions.student_loan_how_many_courses"))
+
+    choose "1"
+    click_on "Continue"
+
+    expect(claim.reload.student_loan_courses).to eql("one_course")
     # TODO [PAGE 33] - When did the first year of your higher education course start
     # TODO [PAGE 34] - When did your higher education courses start
     # [PAGE 35] - Did you take out a postgraduate masters loan on or after 1 August 2016