+++ CHANGELOG.md
@@ -35,6 +35,8 @@ The format is based on [Keep a Changelog]
     - Add 'You are not eligible for an early careers payment'
     - Add 'Are you currently paying off your student loan?'
     - Add 'When you applied for your student loan where was your home address?'
+    - Add 'How many higher education courses did you take a student loan out
+      for?'
   - Personal Details Journey
     - Add 'What is your address?'
     - Add 'What gender does your school's payroll system associate with you?'